chore(gatekeeper): update function to validate upgrade
